fix: resolve lint violations introduced by merging sponsorship-XLS-68-local
- Reformat calculateSponsorFee signature to satisfy prettier's line length.
- Extract SignedAmount into its own file (signed-amount.ts) to bring
  amount.ts back under the max-lines limit, matching the existing
  one-class-per-file convention in types/.
- Suppress prefer-nullish-coalescing on the modify-flag OR chain in
  sponsorshipSet.ts: these are optional booleans, and ?? would stop at an
  explicit false instead of checking the remaining flags.
- Suppress max-statements on the SponsorshipSet test suite, which now
  covers many more scenarios.
