+++ packages/ripple-binary-codec/src/types/amount.ts
@@ -374,61 +374,4 @@ class Amount extends SerializedType {
   }
 }
 
-/**
- * SignedAmount serializes a native XRP magnitude that may be negative.
- *
- * Regular native XRP amounts (Fee, Amount, TakerPays, etc.) can never be
- * negative, so `Amount` continues to reject negative values for those
- * fields. SignedAmount exists only for a small set of fields (such as
- * FeeAmountDelta) whose value is a signed delta -- rippled's STAmount
- * permits a negative native magnitude for these, encoded with the same
- * bit layout as a regular native amount except the "positive" bit (0x40 in
- * the leading byte) is left clear for negative values. Decoding is
- * unaffected and inherited from Amount, since Amount.toJSON() already
- * interprets that bit correctly regardless of which class parsed the bytes.
- */
-class SignedAmount extends Amount {
-  static from<T extends Amount | AmountObject | string>(value: T): SignedAmount {
-    if (value instanceof SignedAmount) {
-      return value
-    }
-
-    if (typeof value !== 'string') {
-      throw new Error('SignedAmount only supports native XRP string values')
-    }
-
-    if (value.indexOf('.') !== -1) {
-      throw new Error(`${value} is an illegal amount`)
-    }
-
-    let decimal: BigNumber
-    try {
-      decimal = new BigNumber(value)
-    } catch (_err) {
-      throw new Error(`${value} is an illegal amount`)
-    }
-    if (!decimal.isZero()) {
-      const magnitude = decimal.abs()
-      if (magnitude.lt(MIN_XRP) || magnitude.gt(MAX_DROPS)) {
-        throw new Error(`${value} is an illegal amount`)
-      }
-    }
-
-    const number = BigInt(value)
-    const isNegative = number < BigInt(0)
-    const magnitude = isNegative ? -number : number
-
-    const intBuf = [new Uint8Array(4), new Uint8Array(4)]
-    writeUInt32BE(intBuf[0], Number(magnitude >> BigInt(32)), 0)
-    writeUInt32BE(intBuf[1], Number(magnitude & BigInt(mask)), 0)
-
-    const amount = concat(intBuf)
-    if (!isNegative) {
-      amount[0] |= 0x40
-    }
-
-    return new SignedAmount(amount)
-  }
-}
-
-export { Amount, SignedAmount, AmountObject }
+export { Amount, AmountObject }
