+++ packages/ripple-binary-codec/src/types/index.ts
@@ -1,5 +1,5 @@
 import { AccountID } from './account-id'
-import { Amount, SignedAmount } from './amount'
+import { Amount } from './amount'
 import { Blob } from './blob'
 import { Currency } from './currency'
 import { Hash128 } from './hash-128'
@@ -11,6 +11,7 @@ import { Issue } from './issue'
 import { STNumber } from './st-number'
 import { PathSet } from './path-set'
 import { STArray } from './st-array'
+import { SignedAmount } from './signed-amount'
 import { STObject } from './st-object'
 import { UInt16 } from './uint-16'
 import { UInt32 } from './uint-32'
